mobileAutoTestsBasicFramework/requirements.txt
Appium-Python-Client>=0.24
pytest==2.9.0
selenium==2.53.1
pytest-testrail>=0.0.11
pytest-xdist>=1.15.0